--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,7 +14,6 @@ readme = "README.md"
 repository = "https://github.com/ssb-ngi-pointer/ssb-uri-rs"
 
 [dependencies]
-anyhow = "1.0.44"
 percent-encoding = "2.1.0"
 regex = "1.5"
 url = "2.2.2"