--- a/src/error.rs
+++ b/src/error.rs
@@ -13,25 +13,25 @@ impl std::fmt::Display for SsbUriError {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match *self {
             SsbUriError::InvalidSuffix(ref err) => {
-                write!(f, "Sigil error: {}", err)
+                write!(f, "{}", err)
             }
             SsbUriError::InvalidUri(ref err) => {
-                write!(f, "URI error: {}", err)
+                write!(f, "{}", err)
             }
             SsbUriError::UnknownFormat(ref err) => {
-                write!(f, "Format error: {}", err)
+                write!(f, "{}", err)
             }
             SsbUriError::UnknownType(ref err) => {
-                write!(f, "Type error: {}", err)
+                write!(f, "{}", err)
             }
             SsbUriError::ParseUrl(ref err) => {
-                write!(f, "Parse error: {}", err)
+                write!(f, "{}", err)
             }
             SsbUriError::InvalidUtf8(ref err) => {
-                write!(f, "Decode error: {}", err)
+                write!(f, "{}", err)
             }
             SsbUriError::InvalidRegex(ref err) => {
-                write!(f, "Regex error: {}", err)
+                write!(f, "{}", err)
             }
         }
     }