This PR removes the [`anyhow`](https://crates.io/crates/anyhow) crate dependency (which provides a trait object based error type) and replaces it with a handwritten, custom error type: `SsbUriError`.

I've been reading about Rust bloat and the factors which influence binary size and compilation time. During my reading, I learned that trait-based error handling libraries - while easy to use - often result in significantly slower compile times. Since ssb-uri-rs is a foundational library, it seems wise to trade a small amount of code complexity for improved compilation times.

**[ before ]**

19 total dependencies

debug compilation   : Finished dev [unoptimized + debuginfo] target(s) in 6.00s
release compilation : Finished release [optimized] target(s) in 8.39s

**[ after ]**

16 total dependencies

debug compilation   : Finished dev [unoptimized + debuginfo] target(s) in 3.42s
release compilation : Finished release [optimized] target(s) in 5.56s

-----

Here's [an article response](https://www.reddit.com/r/rust/comments/gj8inf/rust_structuring_and_handling_errors_in_2020/fqlmknt/) by burntsushi, briefly discussing the benefits of handwritten error impls.